// src/app/models/enums/workspace-routes.ts
export enum WorkspaceRoutes {
  EMPTY = '',
  TRIAL_CASE_ID = ':caseId',
  WORKBOOK = 'workbook',
  CASE_TIMELINE = 'timeline',
  CASE_NARRATIVE = 'narrative',
  TRIAL_NOTEBOOK = 'notebook',
  CASE_TASKS = 'tasks',
  CASE_DOCUMENTS = 'documents',
  CASE_PEOPLE = 'people',
  FOLDER_ID_SECTION_ID = ':folderId/:sectionId',
}